Porting JaegerUdpBatcher tests to JaegerExporterTests.
